@@ -1,12 +1,11 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 28
-    buildToolsVersion '28.0.3'
+    compileSdkVersion 29
     defaultConfig {
         applicationId "man.animalize.ngdaypic"
         minSdkVersion 14
-        targetSdkVersion 28
+        targetSdkVersion 29
         versionCode 1
         versionName "1.0"
     }
@@ -35,6 +34,6 @@ dependencies {
 
     implementation  fileTree(include: ['*.jar'], dir: 'libs')
     implementation  'com.github.bumptech.glide:glide:3.8.0'
-    implementation  'com.android.support:support-v4:28.0.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     implementation  'com.squareup.okhttp3:okhttp:3.12.1'
 }
