@@ -1,9 +1,10 @@
 package man.animalize.ngdaypic;
 
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentActivity;
-import android.support.v4.app.FragmentManager;
+
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentActivity;
+import androidx.fragment.app.FragmentManager;
 
 // 只有一个Fragment的Activity
 public abstract class OneFragmentActivity extends FragmentActivity {
