@@ -1,7 +1,8 @@
 package man.animalize.ngdaypic;
 
 import android.content.Intent;
-import android.support.v4.app.Fragment;
+
+import androidx.fragment.app.Fragment;
 
 public class TouchImageActivity extends OneFragmentActivity {
 
